requirements.txt
python-dotenv
vocode